[sohyun][12-02 PM:03:24]ban com

+++ WebContent/admin/adminMemberList.jsp
@@ -65,7 +65,11 @@
             <script>
             	$("#banIp${list.id}").on("click",function(){
             		var reason = prompt("차단 사유 : ");
-            		location.href="${pageContext.request.contextPath}/admin/adminDeleteMember.admin?id=${list.id}&ip=${list.ipAddr}&reason="+reason;
+            		if(reason == ""){
+            			alert("차단사유는 필수 입력 사항입니다.")
+            		}else{
+                		location.href="${pageContext.request.contextPath}/admin/adminDeleteMember.admin?id=${list.id}&ip=${list.ipAddr}&reason="+reason;
+            		}
             	})
             </script>                
             </c:forEach>
